+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from "./components";
 export { ConnectionConfig } from "./ConnectionConfig";
-export { getOptionsWithDefaults } from "./utils";
 export * from "./constants";
 export * from "./types";