+++ src/utils.ts
@@ -12,16 +12,20 @@ import {
  * with `authenticationType` set to `GoogleAuthType.JWT`. Otherwise, it returns the original options object.
  *
  */
-export function getOptionsWithDefaults<
-  TJsonData extends DataSourceOptions,
-  TSecureJsonData extends DataSourceSecureJsonData
->(options: DataSourceSettings<TJsonData, TSecureJsonData>) {
+export const getOptionsWithDefaults = (
+  options: DataSourceSettings<DataSourceOptions, DataSourceSecureJsonData>,
+  onOptionsChange: (
+    options: DataSourceSettings<DataSourceOptions, DataSourceSecureJsonData>
+  ) => void
+) => {
   if (!options.jsonData.authenticationType) {
-    return {
+    const newOptions = {
       ...options,
       jsonData: { ...options.jsonData, authenticationType: GoogleAuthType.JWT },
     };
+    onOptionsChange(newOptions);
+    return newOptions;
   }
 
   return options;
-}
+};